src/identity_storage/pyproject.toml
[project]
name = "identity-storage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "qdrant-client>=1.14.2",
    "numpy>=1.24.0",
    "pydantic>=1.10.8",
    "uvicorn>=0.34.2",
    "fastapi>=0.115.12",
    "python-dotenv>=1.1.0",
]
